+++ Client/Frontend/Browser/TabManager.swift
@@ -986,98 +986,6 @@ extension TabManager: WKNavigationDelegate {
     }
 }
 
-// WKNavigationDelegates must implement NSObjectProtocol
-class TabManagerNavDelegate: NSObject, WKNavigationDelegate {
-    fileprivate var delegates = WeakList<WKNavigationDelegate>()
-
-    func insert(_ delegate: WKNavigationDelegate) {
-        delegates.insert(delegate)
-    }
-
-    func webView(_ webView: WKWebView, didCommit navigation: WKNavigation!) {
-        for delegate in delegates {
-            delegate.webView?(webView, didCommit: navigation)
-        }
-    }
-
-    func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: Error) {
-        for delegate in delegates {
-            delegate.webView?(webView, didFail: navigation, withError: error)
-        }
-    }
-
-    func webView(_ webView: WKWebView, didFailProvisionalNavigation navigation: WKNavigation!, withError error: Error) {
-        for delegate in delegates {
-            delegate.webView?(webView, didFailProvisionalNavigation: navigation, withError: error)
-        }
-    }
-
-    func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
-        for delegate in delegates {
-            delegate.webView?(webView, didFinish: navigation)
-        }
-    }
-
-    func webViewWebContentProcessDidTerminate(_ webView: WKWebView) {
-        for delegate in delegates {
-            delegate.webViewWebContentProcessDidTerminate?(webView)
-        }
-    }
-
-    func webView(_ webView: WKWebView, didReceive challenge: URLAuthenticationChallenge, completionHandler: @escaping (URLSession.AuthChallengeDisposition, URLCredential?) -> Void) {
-        let authenticatingDelegates = delegates.filter { wv in
-            return wv.responds(to: #selector(webView(_:didReceive:completionHandler:)))
-        }
-
-        guard let firstAuthenticatingDelegate = authenticatingDelegates.first else {
-            return completionHandler(.performDefaultHandling, nil)
-        }
-
-        firstAuthenticatingDelegate.webView?(webView, didReceive: challenge) { (disposition, credential) in
-            completionHandler(disposition, credential)
-        }
-    }
-
-    func webView(_ webView: WKWebView, didReceiveServerRedirectForProvisionalNavigation navigation: WKNavigation!) {
-        for delegate in delegates {
-            delegate.webView?(webView, didReceiveServerRedirectForProvisionalNavigation: navigation)
-        }
-    }
-
-    func webView(_ webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
-        for delegate in delegates {
-            delegate.webView?(webView, didStartProvisionalNavigation: navigation)
-        }
-    }
-
-    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
-        var res = WKNavigationActionPolicy.allow
-        for delegate in delegates {
-            delegate.webView?(webView, decidePolicyFor: navigationAction, decisionHandler: { policy in
-                if policy == .cancel {
-                    res = policy
-                }
-            })
-        }
-        decisionHandler(res)
-    }
-
-    func webView(_ webView: WKWebView,
-                 decidePolicyFor navigationResponse: WKNavigationResponse,
-                 decisionHandler: @escaping (WKNavigationResponsePolicy) -> Void) {
-        var res = WKNavigationResponsePolicy.allow
-        for delegate in delegates {
-            delegate.webView?(webView, decidePolicyFor: navigationResponse, decisionHandler: { policy in
-                if policy == .cancel {
-                    res = policy
-                }
-            })
-        }
-
-        decisionHandler(res)
-    }
-}
-
 // Helper functions for test cases
 extension TabManager {
     func testTabCountOnDisk() -> Int {
